+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.6.5
 
+- The row number of a selected cell now has the same color as the selected cell.
 - bugfix: Updated height calculation for lists 
 
 ## 0.6.4